gateware.iostream: remove in favor of new `IOStreamer`.
