+++ software/glasgow/gateware/jtag/probe.py
@@ -6,7 +6,7 @@ from amaranth.lib import enum, data, wiring, stream, io, cdc
 from amaranth.lib.wiring import In, Out, connect, flipped
 
 from glasgow.gateware.ports import PortGroup
-from glasgow.gateware.iostream2 import IOStreamer
+from glasgow.gateware.iostream import IOStreamer
 
 
 __all__ = ["Mode", "Controller"]
